% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withinsubjerror.R
\name{normDataWithin}
\alias{normDataWithin}
\title{Norms the data within specified groups in a data frame; it normalizes each
subject (identified by idvar) so that they have the same mean, within each group
specified by betweenvars.}
\usage{
normDataWithin(data = NULL, idvar, measurevar, betweenvars = NULL,
  na.rm = FALSE, .drop = TRUE)
}
\description{
data: a data frame.
  idvar: the name of a column that identifies each subject (or matched subjects)
  measurevar: the name of a column that contains the variable to be summariezed
  betweenvars: a vector containing names of columns that are between-subjects variables
  na.rm: a boolean that indicates whether to ignore NA's
}

