% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withinsubjerror.R
\name{summarySE}
\alias{summarySE}
\title{Summarizes data.}
\usage{
summarySE(data = NULL, measurevar, groupvars = NULL, na.rm = FALSE,
  conf.interval = 0.95, .drop = TRUE)
}
\description{
Gives count, mean, standard deviation, standard error of the mean, and confidence interval (default 95%).
  data: a data frame.
  measurevar: the name of a column that contains the variable to be summariezed
  groupvars: a vector containing names of columns that contain grouping variables
  na.rm: a boolean that indicates whether to ignore NA's
  conf.interval: the percent range of the confidence interval (default is 95%)
}

