% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withinsubjerror.R
\name{summarySEwithin}
\alias{summarySEwithin}
\title{Summarizes data, handling within-subjects variables by removing inter-subject variability.}
\usage{
summarySEwithin(data = NULL, measurevar, betweenvars = NULL,
  withinvars = NULL, idvar = NULL, na.rm = FALSE, conf.interval = 0.95,
  .drop = TRUE)
}
\description{
It will still work if there are no within-S variables.
Gives count, mean, standard deviation, standard error of the mean, and confidence interval (default 95%).
If there are within-subject variables, calculate adjusted values using method from Morey (2008).
  data: a data frame.
  measurevar: the name of a column that contains the variable to be summariezed
  betweenvars: a vector containing names of columns that are between-subjects variables
  withinvars: a vector containing names of columns that are within-subjects variables
  idvar: the name of a column that identifies each subject (or matched subjects)
  na.rm: a boolean that indicates whether to ignore NA's
  conf.interval: the percent range of the confidence interval (default is 95%)
}

